Don't run checkInteractNearby on every tick (#1443)
Whatever this method is required for (something with vehicle interaction on mobile I guess) it surely doesn't have to run on every tick. Because of the usage of `getEntityPlayerLookingAt` not running this on every tick will improve the server performance especially when there are many players online. I didn't find any side effects from this change.